User_Service/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.31
Flask-Cors==4.0.1
Werkzeug==3.0.4
Flask-SocketIO
redis
flask_jwt_extended==4.6.0
psycopg2-binary
flask_pydantic_spec
prometheus_flask_exporter
PyJWT==2.9.0
redis-py-cluster
setuptools